requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
cloudpickle==3.0.0
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
dlib==19.24.2
dm-tree==0.1.8
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.50.0
fsspec==2024.3.1
gast==0.4.0
google-auth==2.28.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
imageio==2.19.2
imgaug==0.4.0
importlib_metadata==7.0.2
importlib_resources==6.3.1
jax==0.4.25
Jinja2==3.1.2
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
KNN-NanoFLANN==0.0.0
lazy_loader==0.3
libclang==18.1.1
lmdb==1.3.0
lxml==4.9.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
ml-dtypes==0.3.2
mpmath==1.3.0
networkx==3.2.1
normalSpeed==0.0.1
numpy==1.23.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
onnx==1.12.0
-e git+https://github.com/onnx/onnx-tensorflow.git@ee0c5e537b3cebbddc5773871e6786e6468c7f3f#egg=onnx_tf
onnxruntime==1.9.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==24.0
pandas==1.3.4
Pillow==9.1.1
protobuf==3.20.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-gflags==3.1.2
pytz==2024.1
PyWavelets==1.5.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
safetensors==0.4.2
scikit-image==0.20.0
scipy==1.9.1
shapely==2.0.3
six==1.16.0
sympy==1.12
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow==2.11.0
tensorflow-addons==0.18.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-probability==0.16.0
termcolor==2.4.0
tf-estimator-nightly==2.8.0.dev2021122109
tflite-runtime==2.11.0
tifffile==2024.2.12
timm==0.6.7
torch==2.2.1+cu118
-f https://download.pytorch.org/whl/cu118/torch_stable.html
torchaudio==2.2.1+cu118 
torchvision==0.17.1+cu118
tqdm==4.66.2
triton==2.2.0
typeguard==2.13.3
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.18.1
